+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* x.x.x
+  - Allow reduction functions on the DataContainer class to accept a direction argument which match values in dimension_labels
 
 * 23.1.0
   - Fix bug in IndicatorBox proximal_conjugate